+++ taskmonitor/management/commands/taskmonitorctl.py
@@ -6,7 +6,7 @@ import humanize
 from django.core.management.base import BaseCommand, CommandError
 
 from taskmonitor import __title__, app_settings
-from taskmonitor.core import celery_queues
+from taskmonitor.core import cached_reports, celery_queues
 from taskmonitor.models import TaskLog
 
 CACHE_TIMEOUT_SECONDS = 3600
@@ -34,6 +34,11 @@ class Target(str, Enum):
     SETTINGS = "settings"
 
 
+def my_input(*args, **kwargs) -> str:
+    """Helper to enable mocking of input for unit tests."""
+    return input(*args, **kwargs)
+
+
 class Command(BaseCommand):
     help = f"Command utility for {__title__}."
 
@@ -56,17 +61,31 @@ class Command(BaseCommand):
             title="commands",
             help="available commands",
         )
-
+        # purge command
         parser_purge = subparsers.add_parser(
             UserCommand.PURGE.value, help="Purge a target"
         )
         parser_purge.add_argument(
             Token.TARGET.value,
             type=str,
-            choices=[Target.QUEUE.value],
+            choices=[Target.QUEUE.value, Target.LOGS.value],
             help="target to purge",
         )
+        purge_group = parser_purge.add_mutually_exclusive_group(required=True)
+        purge_group.add_argument(
+            "--all", action="store_true", help="You want to purge everything"
+        )
+        purge_group.add_argument(
+            "--app-name", help="Limit purge to tasks from an specific app"
+        )
+        purge_group.add_argument(
+            "--task-id", help="Limit purge to tasks with a specific ID"
+        )
+        purge_group.add_argument(
+            "--task-name", help="Limit purge to tasks with a specific name"
+        )
 
+        # inspect command
         parser_inspect = subparsers.add_parser(
             UserCommand.INSPECT.value,
             help="Inspect a target, e.g. show information about it.",
@@ -83,25 +102,77 @@ class Command(BaseCommand):
             help="Fore re-calculation of values and update caches.",
         )
 
-    def user_confirmed(self, question_text):
-        user_input = input(f"{question_text} (y/N)?")
+    def _user_confirmed(self, question_text):
+        """Ask user about confirmation and exit
+        when he does not reply in the affirmative.
+        """
+        user_input = my_input(f"{question_text} (y/N)?")
         if user_input.lower() != "y":
             self.stdout.write(self.style.WARNING("Aborted by user request."))
             exit(1)
 
-    def purge_queue(self):
+    def _clear_line(self):
+        self.stdout.write("" * 70, ending="\r")
+
+    def purge_queue(self, options):
         num_entries = celery_queues.queue_length()
         if not num_entries:
             self.stdout.write(self.style.WARNING("Queue is empty. Aborted."))
             exit(1)
-        self.user_confirmed(
-            f"Are you sure you purge {num_entries:,} tasks from the queue?"
-        )
-        celery_queues.clear_tasks()
-        self.stdout.write(f"Purged {num_entries:,} tasks from queue...")
+        self.stdout.write(f"Current queue size: {num_entries:,}")
+        if options["task_name"]:
+            task_name = options["task_name"]
+            self._user_confirmed(
+                f"Are you sure you purge all tasks with the TASK NAME {task_name} from the queue?"
+            )
+            self.stdout.write("Purged tasks from queue...", ending="\r")
+            deleted_entries = celery_queues.delete_task_by_name(task_name)
+            self._clear_line()
+        elif options["task_id"]:
+            task_id = options["task_id"]
+            self._user_confirmed(
+                f"Are you sure you purge all tasks with the TASK ID {task_id} from the queue?"
+            )
+            self.stdout.write("Purged tasks from queue...", ending="\r")
+            deleted_entries = celery_queues.delete_task_by_id(task_id)
+            self._clear_line()
+        elif options["app_name"]:
+            app_name = options["app_name"]
+            self._user_confirmed(
+                f"Are you sure you purge all tasks belonging to the app {app_name} from the queue?"
+            )
+            self.stdout.write("Purged tasks from queue...", ending="\r")
+            deleted_entries = celery_queues.delete_task_by_app_name(app_name)
+            self._clear_line()
+        elif options["all"]:
+            self._user_confirmed("Are you sure you purge ALL TASKS from the queue?")
+            self.stdout.write("Purged tasks from queue...", ending="\r")
+            celery_queues.clear_tasks()
+            deleted_entries = num_entries
+            self._clear_line()
+        else:
+            raise NotImplementedError("This option is not yet implemented")
+        self.stdout.write(f"Purged {deleted_entries:,} tasks from queue...")
         self.stdout.write(self.style.SUCCESS("Done."))
 
-    def inspect_logs(self):
+    def purge_logs(self, options):
+        all_logs = TaskLog.objects.all()
+        self.stdout.write("Calculating...", ending="\r")
+        num_logs = all_logs.count()
+        if not num_logs:
+            self.stdout.write(self.style.WARNING("No logs found. Aborted."))
+            exit(1)
+        self.stdout.write(f"Current task logs count: {num_logs:,}")
+        if options["all"]:
+            self._user_confirmed("Are you sure you purge ALL task logs?")
+            self.stdout.write("Deleting task logs...")
+            all_logs._raw_delete(all_logs.db)
+        else:
+            raise CommandError("Currently only the --all option is supported")
+        cached_reports.refresh_cache()
+        self.stdout.write(self.style.SUCCESS("Done."))
+
+    def inspect_logs(self, options):
         log_count = TaskLog.objects.count()
         try:
             db_table_size = TaskLog.objects.db_table_size()
@@ -122,14 +193,31 @@ class Command(BaseCommand):
         for label, value in output.items():
             self.stdout.write(f"{label:{max_length + 1}}: {value}")
 
-    def inspect_queue(self):
+    def inspect_queue(self, options):
         num_entries = celery_queues.queue_length()
+        if not num_entries:
+            self.stdout.write("Queue is empty.")
+            return
         self.stdout.write(f"Current queue size: {num_entries:,}")
-        self.stdout.write("Summary of queued tasks by app count in descending order:")
-        app_in_tasks = (
-            task.app_name for task in celery_queues._fetch_task_from_all_queues()
-        )
-        field_counts = Counter(app_in_tasks)
+        if num_entries > app_settings.TASKMONITOR_QUEUED_TASKS_ADMIN_LIMIT:
+            self._user_confirmed(
+                "The queue is very large. Do you still want to gather statistics?"
+            )
+        self.stdout.write("Fetching data from queue...", ending="\r")
+        self._clear_line()
+        tasks = [
+            (task.app_name, task.name)
+            for task in celery_queues._fetch_task_from_all_queues()
+        ]
+        self.stdout.write("Queued tasks grouped by app counts in descending order:")
+        grouped_names = (o[0] for o in tasks)
+        self._render_field_counts(grouped_names)
+        self.stdout.write("Queued tasks grouped by task counts in descending order:")
+        grouped_names = (o[1] for o in tasks)
+        self._render_field_counts(grouped_names)
+
+    def _render_field_counts(self, grouped_names):
+        field_counts = Counter(grouped_names)
         field_counts_sorted = dict(
             sorted(field_counts.items(), key=lambda item: item[1], reverse=True)
         )
@@ -141,7 +229,7 @@ class Command(BaseCommand):
         for app_name, count in field_counts_sorted.items():
             self.stdout.write(f"  {app_name:{max_length}}: {count:,}")
 
-    def inspect_settings(self):
+    def inspect_settings(self, options):
         settings = sorted(
             [o for o in dir(app_settings) if not o.startswith("__") and o == o.upper()]
         )
@@ -154,21 +242,8 @@ class Command(BaseCommand):
         command = options[Token.COMMAND.value]
         target = options[Token.TARGET.value]
 
-        if command == UserCommand.PURGE:
-            if target == Target.QUEUE:
-                self.purge_queue()
-            else:
-                raise NotImplementedError()
-
-        elif command == UserCommand.INSPECT:
-            if target == Target.QUEUE:
-                self.inspect_queue()
-            elif target == Target.LOGS:
-                self.inspect_logs()
-            elif target == Target.SETTINGS:
-                self.inspect_settings()
-            else:
-                raise NotImplementedError()
-
-        else:
-            raise NotImplementedError()
+        method = f"{command}_{target}"
+        try:
+            getattr(self, method)(options)
+        except AttributeError:
+            raise NotImplementedError(method) from None