+++ taskmonitor/tests/generate_task_logs.py
@@ -15,13 +15,20 @@ os.environ.setdefault("DJANGO_SETTINGS_MODULE", "myauth.settings.local")
 django.setup()
 
 """MAIN"""
+from taskmonitor.core import cached_reports
 from taskmonitor.models import TaskLog
 from taskmonitor.tests.factories import TaskLogFactory
 
-MAX_ENTRIES = 50_000
+MAX_ENTRIES = 500_000
+MAX_BATCH = 10_000
 
-print(f"Generating {MAX_ENTRIES:,} task logs...")
-objs = TaskLogFactory.build_batch(size=MAX_ENTRIES)
-print("Storing...")
-TaskLog.objects.bulk_create(objs, batch_size=500, ignore_conflicts=True)
+assert MAX_BATCH <= MAX_ENTRIES
+assert MAX_ENTRIES % MAX_BATCH == 0
+print(f"Generating a total of {MAX_ENTRIES:,} task logs...")
+max_runs = MAX_ENTRIES // MAX_BATCH
+for run in range(max_runs):
+    print(f"Generating {MAX_BATCH:,} logs - {run + 1:,} / {max_runs:,}")
+    objs = TaskLogFactory.build_batch(size=MAX_BATCH)
+    TaskLog.objects.bulk_create(objs, batch_size=500, ignore_conflicts=True)
+cached_reports.refresh_cache()
 print("DONE!")