+++ taskmonitor/views.py
@@ -9,14 +9,16 @@ from django.shortcuts import redirect, render
 
 from allianceauth import NAME as site_header
 from allianceauth.services.hooks import get_extension_logger
-from app_utils.caching import cached_queryset
 from app_utils.logging import LoggerAddTag
 
-from . import __title__, tasks
-from .app_settings import TASKMONITOR_DATA_MAX_AGE, TASKMONITOR_REPORTS_MAX_TOP
-from .core import cached_reports, celery_queues
-from .helpers import Echo
-from .models import TaskLog
+from taskmonitor import __title__, tasks
+from taskmonitor.app_settings import (
+    TASKMONITOR_DATA_MAX_AGE,
+    TASKMONITOR_REPORTS_MAX_TOP,
+)
+from taskmonitor.core import cached_reports, celery_queues
+from taskmonitor.helpers import Echo
+from taskmonitor.models import TaskLog
 
 logger = LoggerAddTag(get_extension_logger(__name__), __title__)
 
@@ -49,22 +51,10 @@ def admin_taskmonitor_download_csv(request) -> StreamingHttpResponse:
 @staff_member_required
 def admin_taskmonitor_reports(request):
     """Show the reports page."""
-    timeout = 60
-    total_runs = cached_queryset(
-        queryset=TaskLog.objects.count(),
-        key="tasklog-reports-total-runs",
-        timeout=timeout,
-    )
-    oldest_date = cached_queryset(
-        queryset=TaskLog.objects.oldest_date(),
-        key="tasklog-reports-oldest-data",
-        timeout=timeout,
-    )
-    newest_date = cached_queryset(
-        queryset=TaskLog.objects.newest_date(),
-        key="tasklog-reports-newest-date",
-        timeout=timeout,
-    )
+    report = cached_reports.report("tasks_basics").data()
+    total_runs = report["total_runs"]
+    oldest_date = report["oldest_date"]
+    newest_date = report["newest_date"]
     context = {
         "title": "Reports",
         "site_header": site_header,